Complete documentation for DCEL
Changelog:
- Add new datasets to Convex Hull
- Add datasets to DCEL
- Generate Docs for DCEL

+++ ConvexHull/src/PointList.h
@@ -58,7 +58,7 @@ class PointList {
         }
 
         /// To provide interface to STL vector's emplace_back() function
-        auto emplace_back(long a, long b) {
+        auto emplace_back(double a, double b) {
             return pts.emplace_back(a, b);
         }
